Move schemas/utils from auth to the relevant directories
