+++ .env.example
@@ -1,7 +1,14 @@
 # App configuration
-==============================
+# ==============================
 # host of your app (localhost or 127.0.0.1 is default)
-APP_HOST=
+APP_HOST=127.0.0.1
 # port of your app (8000 default)
-APP_PORT=
-==============================
+APP_PORT=8000
+# ==============================
+
+# Auth settings
+# ==============================
+APP_SECRET_KEY=my-cool-secret-key
+APP_ALGORITHM=HS256
+APP_ACCESS_TOKEN_EXPIRE_MINUTES=15
+# ==============================
