+++ app/api/v1/enpoints/auth.py
@@ -1,17 +1,11 @@
-from datetime import UTC, datetime, timedelta
-from uuid import UUID
-
-import jwt
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from jwt.exceptions import InvalidTokenError
-from passlib.context import CryptContext
-from pydantic import BaseModel
 
 from core.config import auth_settings
+from schemas.token import Token
+from schemas.user import User, UserAuth
+from utils.auth import create_jwt, verify_password
 
-# to get a string like this run:
-# openssl rand -hex 32
 SECRET_KEY = auth_settings.secret_key
 ALGORITHM = auth_settings.algorithm
 ACCESS_TOKEN_EXPIRE_MINUTES = auth_settings.access_token_expire_minutes
@@ -32,59 +26,11 @@ fake_users_db = {
 }
 
 
-class User(BaseModel):
-    user_id: UUID
-    user_name: str
-
-
-class UserInDB(User):
-    hashed_password: str
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    user_name: str
-
-
-# get user credentials
 # get user data
-async def get_user(user_name: str) -> UserInDB | None:
+async def get_user(user_name: str) -> UserAuth | None:
     if user_name in fake_users_db:
-        return UserInDB(**fake_users_db[user_name])
+        return UserAuth(**fake_users_db[user_name])
     return None
-
-
-# check user password
-# -> 1. Get hash method
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-# -> 2. Hash password
-def hash_password(password: str):
-    return pwd_context.hash(password)
-
-
-# -> 3. Verify password
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-# generate JWT
-# -> 1. Create new JWT
-def create_jwt(user_data: User, expires_delta: timedelta = timedelta(minutes=DEFAULT_EXPIRE_MINUTES)):
-    to_encode = user_data.dict()
-    expire = datetime.utcnow() + expires_delta
-    to_encode["exp"] = expire
-    # ==============================
-    # uuid is not json serializable
-    to_encode["user_id"] = str(user_data.user_id)
-    # ==============================
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
 
 
 @router.post("/token")
@@ -94,39 +40,19 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
         raise HTTPException(status_code=400, detail="Incorrect username or password")
     if not verify_password(form_data.password, user.hashed_password):
         raise HTTPException(status_code=400, detail="Incorrect username or password")
-    access_token = create_jwt(user)
+    user_without_password = User(**user.model_dump())
+    access_token = create_jwt(user_without_password)
     return Token(access_token=access_token, token_type="bearer")
 
 
-async def get_current_user(token: str = Depends(oauth2_scheme)):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        user_name = payload.get("sub")
-        if user_name is None:
-            raise credentials_exception
-        token_data = TokenData(user_name=user_name)
-    except InvalidTokenError:
-        raise credentials_exception from InvalidTokenError
-    user = get_user(user_name=token_data.user_name)
-    if user is None:
-        raise credentials_exception
-    return user
+# registration:
+# -> 1. Get user credentionals.
+# -> 2. Check that not user with this login/email/phone/etc.
+# -> 3. Make sure that password is complex enough.
+# -> 4. Save user creds in the DB (with hashed password).
+# -> 5. Autorize user with this creds.
 
 
-@router.get("/me")
-async def get_current_active_user(
-    current_user: User = Depends(get_current_user),
-):
-    if not current_user:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-
-@router.get("Hello world!")
-async def hello_world():
-    return "hello world!"
+@router.get("/protected")
+async def protected(protected=Depends(oauth2_scheme)):
+    return "hello!"
